% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishkirkko2015.R
\docType{data}
\name{fishkirkkojarvi2015}
\alias{fishkirkkojarvi2015}
\title{302 measurements of fish species at Kirkkojarvi lake}
\format{dataframe with 302 rows and 6 columns:
\describe{
 \item{fishname}{Fish name in Finnish}
 \item{fishID}{Fish unique identifier for this dataset}
 \item{sl}{Standard Length in mm}
 \item{fl}{Fork Length in mm}
 \item{tl}{Total Length in mm}
 \item{wt}{Weight in g}
}}
\usage{
fishkirkkojarvi2015
}
\description{
302 lenght measurements of 11 fish species at Kirkkojarvi lake
}
\examples{
\dontrun{
# get an histogram of the fish species in the dataset
data(fishkirkkojarvi2015)
hist(fishkirkkojarvi2015[["fishID"]], xaxt="n", 
main='Histogram of fish species from the Kirkkojarvi lake',xlab='Fish species')
axis(1,at=1:10,labels=fishnames[1:10,'English'])
}
}
\author{
Jose Gama
}
\keyword{data}

