% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishkirkko2015.R
\docType{data}
\name{fishnames}
\alias{fishnames}
\title{Names of fish species from Kirkkojarvi lake in four languages}
\format{dataframe with 11 rows and 5 columns:
\describe{
 \item{fishID}{fish unique identifier for this dataset}
 \item{Finnish}{Fish name in Finnish}
 \item{English}{Fish name in English}
 \item{binomial.name}{Fish binomial name in Latin}
 \item{Swedish}{Fish name in Swedish}
}}
\usage{
fishnames
}
\description{
Names of fish species from Kirkkojarvi lake in Finnish, 
Swedish, English and Latin (binomial name)
}
\examples{
# What is "Kuha" in English and its binomial name?
data(fishnames)
fishnames[which(fishnames[["Finnish"]]=='Kuha'),c('English','binomial.name')]
}
\author{
Jose Gama
}
\keyword{data}

